Move localStorage availability check to RemoteStorage.util
