+++ src/localstorage.js
@@ -90,15 +90,7 @@
   RemoteStorage.LocalStorage._rs_init = function () {};
 
   RemoteStorage.LocalStorage._rs_supported = function () {
-    if (!('localStorage' in global)) { return false }
-
-    try {
-      global.localStorage.setItem('rs-check', 1);
-      global.localStorage.removeItem('rs-check');
-      return true;
-    } catch(error) {
-      return false;
-    }
+    return RemoteStorage.util.localStorageAvailable();
   };
 
   // TODO tests missing!
