+++ src/util.js
@@ -401,8 +401,22 @@
       }
 
       return md5(str);
-    }
+    },
     /* jshint ignore:end */
+
+
+    localStorageAvailable: function() {
+      if (!('localStorage' in global)) { return false }
+
+      try {
+        global.localStorage.setItem('rs-check', 1);
+        global.localStorage.removeItem('rs-check');
+        return true;
+      } catch(error) {
+        return false;
+      }
+    }
+
   };
 
   if (!RemoteStorage.prototype.util) {
